Added keyboard controller

+++ AsteroidsGame/Sprites/Point2D.cs
@@ -7,19 +7,19 @@ namespace AsteroidsGame.Sprites
 {
     public class Vector
     {
-        public int x { get; set; }
-        public int y { get; set; }
+        public float x { get; set; }
+        public float y { get; set; }
 
         public Vector():this(0,0)
         {
         }
-        public Vector(int x, int y)
+        public Vector(float x, float y)
         {
             this.x = x;
             this.y = y;
         }
 
-        public void Move(int x,int y)
+        public void Move(float x,float y)
         {
             this.x +=  x;
             this.y +=  y;
